+++ docs/CHANGELOG.md
@@ -1,14 +1,14 @@
 # GoNB Changelog
 
-## Next
+## 0.7.1 - 2023/07/03
 
 * Added support for tracking `go.work`, which allows auto-complete and contextual help
-  to work with the local modules configured. It also requires `gopls` _v0.12.4_ or newer to work.
+  to work with the local modules configured. It also requires `gopls` **v0.12.4** or newer to work.
 * Fixed auto-complete bug when no `main` function (or no `%%`) was present in cell.
 * Added special command `%cd` to chance current directory.
 * Commands `%cd` and `%env` prints results of its execution.
 
-## v0.70.0 - 2023/05/29
+## v0.7.0 - 2023/05/29
 
 * Added "%ls" and "%rm" to manage memorized definitions directly.
 * More contextual help and auto-complete improvements: